create an interface to be injected into the league service, this prepares us for eventually swapping to complete nhost implementation
