+++ src/repositories/league-data.repository.ts
@@ -1,0 +1,53 @@
+import { GuildMember } from "discord.js";
+import { LeaguePlayer, SheetsPlayer } from "../models/league-models";
+
+export interface SignupData {
+  teamName: string;
+  teamNoDays: string;
+  teamCompKnowledge: string;
+  player1: SheetsPlayer;
+  player2: SheetsPlayer;
+  player3: SheetsPlayer;
+  additionalComments: string;
+}
+
+export interface SignupResult {
+  rowNumber: number | null;
+  seasonInfo: {
+    signupPrioEndDate: string;
+    startDate: string;
+  };
+}
+
+export interface SubRequestData {
+  teamDivision: string;
+  teamName: string;
+  weekNumber: string;
+  playerOut: LeaguePlayer;
+  playerIn: LeaguePlayer;
+  playerInDivision: string;
+  commandUser: GuildMember;
+  additionalComments: string;
+}
+
+export interface RosterChangeData {
+  teamDivision: string;
+  teamName: string;
+  playerOut: LeaguePlayer;
+  playerIn: LeaguePlayer;
+  commandUser: GuildMember;
+  additionalComments: string;
+}
+
+export interface WriteResult {
+  rowNumber: number | null;
+  url: string;
+  tabName: string;
+}
+
+export interface LeagueDataRepository {
+  getRosterDiscordIds(): Promise<Map<string, string>>;
+  writeSignup(data: SignupData): Promise<SignupResult | null>;
+  writeSubRequest(data: SubRequestData): Promise<WriteResult>;
+  writeRosterChange(data: RosterChangeData): Promise<WriteResult>;
+}
